Allow specifying a separate label for Key in keyvaluelist inputs
`label` can now accept either a `string`, as before, or a more complex
object of `{key: string; value: string}` to allow defining labels for
both labels in `keyvaluelist` collection inputs.

`key` is only passed on to the server if `collection` is set to `keyvaluelist`
though it will accept both forms - `key` is dropped at submission-time.