updates convertOnExecution and convertComponentReference to pass along default values
Also updates default values from templates so that those can be consumed by the convert layer.

The interator was changed in both conversion functions to iterate over the inputs. Those inputs have all of the keys and potential defaults. We then use the provided values from the user to override any values that are present and pass along the defaults if they arent.

Bumps version to 9.0.7