Run input clean functions even if value not explicitly provided
Previous behavior was to iterate over the collection of provided inputs
but this meant that certain scenarios, such as version upgrades, wouldn't
get an opportunity to run input clean functions if the value wasn't present
in the collection of input values.

Now clean functions will be given `undefined` and an opportunity to run
and clean the value into something usable/expected.